libcamera: fix bug in exposure compensation functionality
BZ: 1309
wrong init parameters for exposure compensation and step be not considered

make ev step 1/3, ev index range -6~+6, and use float type for EV bias
fix bug in libcamera for float EV bias support

+++ CameraAAAProcess.h
@@ -146,8 +146,8 @@ public:
     int AeSetMeteringMode(ci_adv_AeMeteringMode mode);
     int AeGetMeteringMode(ci_adv_AeMeteringMode *mode);
 
-    int AeSetEv(int bias);
-    int AeGetEv(int *bias);
+    int AeSetEv(float bias);
+    int AeGetEv(float *bias);
 
     int AeLock(bool lock) {
         return ci_adv_AeLock(lock);
